Changed back to integer data type and modified js file

+++ app/database/migrations/2014_11_25_055745_create_canvas_entities_table.php
@@ -16,11 +16,11 @@ class CreateCanvasEntitiesTable extends Migration {
 		{
 			$table->increments('id');
 			$table->string('name');
-			$table->decimal('x');
-			$table->decimal('y');
-			$table->decimal('dx');
-			$table->decimal('dy');
-			$table->decimal('radius');
+			$table->integer('x');
+			$table->integer('y');
+			$table->integer('dx');
+			$table->integer('dy');
+			$table->integer('radius');
 			$table->timestamps();
 		});
 	}
